requirements.txt
robotframework
robotframework-jsonlibrary
robotframework-browser
robotframework-requests
pyyaml
PyJWT
oscar-python